% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/same_pop_test.R
\name{same_pop_test}
\alias{same_pop_test}
\title{A simple falsification test for the Two-Sample Mendelian randomisation 'same population' assumption.}
\usage{
same_pop_test(
  Bexp,
  Bout,
  SEexp,
  SEout,
  SNPlist,
  Fisher = F,
  Random = T,
  Fixed = T,
  sm = "md"
)
}
\arguments{
\item{Bexp}{SNP-phenotype associations for the exposure sample.}

\item{Bout}{SNP-phenotype associations for the outcome sample.}

\item{SEexp}{Standard error of the SNP-phenotype associations for the exposure sample.}

\item{SEout}{Standard error of the SNP-phenotype associations for the outcome sample.}

\item{SNPlist}{list of rsid's for the SNPs used.}

\item{Fisher}{Logical indicating if to Fisher's methods (meta-analysis of p-vlalues) to combine differences}

\item{Random}{Logical indicating if a random effects meta-analysis should run, defult = T.}

\item{Fixed}{Logical indicating if a fiexed effects meta-analysis should run, defult = T.}

\item{sm}{A character string indicating underlying summary measure, e.g., "RD", "RR", "OR", "ASD", "HR", "MD", "SMD", or "ROM".}
}
\description{
This function is a falsification test for the 'same population' assumption made by Two-Sample Mendelian randomisation. Please note that this function tests the absolute difference in SNP effects such that negative effects imply that the exposure GWAS has smaller betas than the outcome GWAS
}
\examples{
library(TwoSampleMR)
#extracting list of SNPs used as instruments
exp <- extract_instruments(outcomes="ieu-b-4760")
#extracting data on outcome from outcome sample
out<- extract_outcome_data( exp$SNP,  c("ieu-a-1009"))
#extracting data on outcome from exposure sample
ukb_out<-extract_outcome_data(dat$SNP,"ukb-b-4062")
#combining data for analysis
ukb_out$beta.ukb<-ukb_out$beta.outcome
ukb_out$se.ukb<-ukb_out$se.outcome
ukb_out<-ukb_out[,c("SNP","beta.ukb","se.ukb","effect_allele.outcome")]
snp_comp<-merge(dat[,c("SNP","beta.outcome","effect_allele.exposure","se.outcome")],ukb_out, by="SNP")
#running analysis
same_pop_test(SNPlist=snp_comp$SNP, Bexp=snp_comp$beta.ukb, Bout=snp_comp$beta.outcome, SEexp=snp_comp$se.ukb, SEout=snp_comp$se.outcome, Random=T, Fixed=F)
#example taken from (insert DIO)
}
\keyword{2SMR}
